Fix svcat create class (#2392)
* Disallow all scope for create class

* Copy just the spec when creating a new class

The test was passing because key data on the object metadata wasn't set
in the test, but when you run the command directly in svcat, it failed
due to ResourceVersion being set.

* Add failing test for svcat create class

* sync golden files

* Set the external name to the new class name

svcat displays the external name, because the k8s name is usually an unfriendly
generated id. By setting the external name as well, svcat will show the new name
selected by the user.